+++ certificate/authorization/doc.go
